requirements.txt
slack-bolt
slack-sdk
anthropic
pdfplumber
python-dotenv
